requirements.txt
Django==4.2.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.3.*
python-dotenv==1.0.*
requests==2.32.*
requests-oauthlib==2.0.*
django-environ==0.11.*
psycopg2-binary==2.9.*
gunicorn==21.2.*
django-cors-headers>=4.0.0
openai==1.12.*
Pillow==10.2.*
httpx>=0.27.0  # HTTP client for OpenRouter streaming
whitenoise==6.6.0
fal-client>=0.10.0
mixpanel==5.0.0
sentry-sdk[django]>=2.0.0
cairosvg>=2.7.0  # SVG to PNG conversion for logo processing
boto3>=1.34.0  # S3-compatible storage (Cloudflare R2) for public media URLs
celery>=5.3.0  # Async task queue for background processing
redis>=5.0.0  # Redis client for Celery broker
reportlab>=4.0.0  # PDF generation
python-pptx>=0.6.21  # PPTX generation
requests>=2.32.*  # Already present, used for image downloads
cryptography>=41.0.0  # Encryption for secrets
postgrest>=0.13.0  # Supabase PostgREST client
jsonschema>=4.20.0  # JSON schema validation
python-jsonschema-objects>=0.9.0  # JSON schema objects
esbuild>=0.19.0  # Bundle compilation (via subprocess)
postgrest>=0.13.0  # Supabase PostgREST client
supabase>=2.0.0  # Supabase Python client
PyMySQL>=1.1.0  # MySQL database connector
SQLAlchemy>=2.0.0  # SQL toolkit for database introspection

